dehy/requirements.txt
asgiref==3.5.0
astroid==2.9.3
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
beautifulsoup4==4.10.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
Django==3.2.11
django-better-admin-arrayfield==1.4.2
django-environ==0.8.1
django-extra-views==0.13.0
django-haystack==3.1.1
django-oscar==3.1
django-phonenumber-field==5.2.0
django-sslserver==0.22
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==11.3.0
flake8==4.0.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.2.0
isort==5.10.1
jedi==0.17.2
jsonfield==3.1.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.21.5
openpyxl==3.0.9
outcome==1.1.0
pandas==1.3.5
parso==0.7.1
phonenumbers==8.12.41
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.3
purl==1.6
pycodestyle==2.8.0
pycountry==22.1.10
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
pygit2==1.7.2
pylint==2.12.2
pyls-mypy==0.1.8
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
pytz==2021.3
requests==2.27.1
rope==0.22.0
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sorl-thumbnail==12.6.3
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
stripe==2.65.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.0
trio==0.20.0
trio-websocket==0.9.2
typed-ast==1.5.2
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.8
wrapt==1.13.3
wsproto==1.1.0
yapf==0.32.0
zipp==3.7.0